+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Mathematica for sums of powers in Eulerian numbers
 - Update conclusions
 - Add formula for ordinary sums of powers
+- Add slides
 
 ## [1.3.0] - 03-Jan-2026